refactor: move sunburst to shared and pass taxa tree per site (#1598)
