+++ app/components/Home/components/Section/components/SectionAssemblies/sectionAssemblies.tsx
@@ -1,8 +1,10 @@
-import { SectionViz as Sunburst } from "@/components/Home/components/Section/components/SectionViz/sunburst";
+import { Sunburst } from "@repo/shared/views/HomeView/components/Section/components/SectionAssemblies/components/Sunburst/sunburst";
+import { type TaxonomyNode } from "@repo/shared/views/HomeView/components/Section/components/SectionAssemblies/components/Sunburst/types";
 import {
   SectionSubtitle,
   SectionTitle,
 } from "@repo/shared/views/HomeView/components/Section/section.styles";
+import taxaTree from "catalog/output/ncbi-taxa-tree.json";
 import { type JSX } from "react";
 import { Headline, Section, SectionLayout } from "./sectionAssemblies.styles";
 
@@ -16,7 +18,7 @@ export const SectionAssemblies = (): JSX.Element => {
             Browse assemblies by taxonomic lineage.
           </SectionSubtitle>
         </Headline>
-        <Sunburst />
+        <Sunburst data={taxaTree as TaxonomyNode} logoPath="/logo/brc.svg" />
       </SectionLayout>
     </Section>
   );
