--- app/theme/color-maps/tol.ts
@@ -1,67 +1,0 @@
-// https://sronpersonalpages.nl/~pault/
-export const bright = [
-  "#4477AA",
-  "#EE6677",
-  "#228833",
-  "#CCBB44",
-  "#66CCEE",
-  "#AA3377",
-  "#BBBBBB",
-];
-export const high_contrast = ["#004488", "#DDAA33", "#BB5566"];
-export const vibrant = [
-  "#EE7733",
-  "#0077BB",
-  "#33BBEE",
-  "#EE3377",
-  "#CC3311",
-  "#009988",
-  "#BBBBBB",
-];
-export const muted = [
-  "#CC6677",
-  "#332288",
-  "#DDCC77",
-  "#117733",
-  "#88CCEE",
-  "#882255",
-  "#44AA99",
-  "#999933",
-  "#AA4499",
-  "#DDDDDD",
-];
-export const medium_contrast = [
-  "#6699CC",
-  "#004488",
-  "#EECC66",
-  "#994455",
-  "#997700",
-  "#EE99AA",
-];
-export const light = [
-  "#77AADD",
-  "#EE8866",
-  "#EEDD88",
-  "#FFAABB",
-  "#99DDFF",
-  "#44BB99",
-  "#BBCC33",
-  "#AAAA00",
-  "#DDDDDD",
-];
-export const dark = [
-  "#222255",
-  "#225555",
-  "#225522",
-  "#666633",
-  "#663333",
-  "#555555",
-];
-export const pale = [
-  "#BBCCEE",
-  "#CCEEFF",
-  "#CCDDAA",
-  "#EEEEBB",
-  "#FFCCCC",
-  "#DDDDDD",
-];
