+++ packages/shared/views/HomeView/components/Section/components/SectionAssemblies/components/Sunburst/constants.ts
@@ -1,0 +1,8 @@
+import { type TaxonomyNode } from "./types";
+
+export const DEFAULT_TAXONOMY_NODE: TaxonomyNode = {
+  assembly_count: 0,
+  children: [],
+  name: "root",
+  ncbi_tax_id: "",
+};
